+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add `interpretability` module for generating similarity maps
+
 ### Changed
 
 - Remove dummy image from `ColPaliProcessor.process_queries`
-
-### Fixed
 
 ## [0.3.1] - 2024-09-27